requirements.txt
functions-framework
google-api-python-client==2.70.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
google-cloud-storage==2.5.0
openai==1.47.0
requests==2.28.1
